Clean-up code
